Improve: add useAdminStorages composables
